# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SLA Pause & Escalation Service",
    "description": "逾期升级 SLA 暂停服务。\n\n- 暂停申请绑定事件 / 原因 / 证据，状态机：申请 → 批准生效 / 拒绝 → 结束 / 撤销；\n- 升级计算基于注入时钟，扣除已批准暂停区间的**并集**（重叠不重复延长）；\n- 结案后时钟冻结，不再升级，暂停不得重新启动；\n- 暂停不删除已产生升级、不改写已锁定处罚；迟到批准只产生待复核补偿建议，\n  确认后追加更正并保留原处罚链。\n",
    "version": "1.0.0"
  },
  "paths": {
    "/cases": {
      "post": {
        "tags": [
          "cases"
        ],
        "summary": "Create Case",
        "operationId": "create_case_cases_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CaseCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cases/{case_id}": {
      "get": {
        "tags": [
          "cases"
        ],
        "summary": "Get Case",
        "operationId": "get_case_cases__case_id__get",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cases/{case_id}/close": {
      "post": {
        "tags": [
          "cases"
        ],
        "summary": "Close Case",
        "operationId": "close_case_cases__case_id__close_post",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cases/{case_id}/pause-requests": {
      "post": {
        "tags": [
          "pauses"
        ],
        "summary": "Apply Pause",
        "operationId": "apply_pause_cases__case_id__pause_requests_post",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Case Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PauseApplyIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PauseOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "pauses"
        ],
        "summary": "List Case Pauses",
        "operationId": "list_case_pauses_cases__case_id__pause_requests_get",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PauseOut"
                  },
                  "title": "Response List Case Pauses Cases  Case Id  Pause Requests Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/pause-requests": {
      "get": {
        "tags": [
          "pauses"
        ],
        "summary": "List Pauses",
        "operationId": "list_pauses_pause_requests_get",
        "parameters": [
          {
            "name": "case_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Case Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/PauseStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PauseOut"
                  },
                  "title": "Response List Pauses Pause Requests Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/pause-requests/{pause_id}": {
      "get": {
        "tags": [
          "pauses"
        ],
        "summary": "Get Pause",
        "operationId": "get_pause_pause_requests__pause_id__get",
        "parameters": [
          {
            "name": "pause_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Pause Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PauseOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/pause-requests/{pause_id}/approve": {
      "post": {
        "tags": [
          "pauses"
        ],
        "summary": "Approve Pause",
        "operationId": "approve_pause_pause_requests__pause_id__approve_post",
        "parameters": [
          {
            "name": "pause_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Pause Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PauseApproveIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PauseOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/pause-requests/{pause_id}/reject": {
      "post": {
        "tags": [
          "pauses"
        ],
        "summary": "Reject Pause",
        "operationId": "reject_pause_pause_requests__pause_id__reject_post",
        "parameters": [
          {
            "name": "pause_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Pause Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PauseRejectIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PauseOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/pause-requests/{pause_id}/end": {
      "post": {
        "tags": [
          "pauses"
        ],
        "summary": "End Pause",
        "operationId": "end_pause_pause_requests__pause_id__end_post",
        "parameters": [
          {
            "name": "pause_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Pause Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PauseEndIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PauseOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/pause-requests/{pause_id}/cancel": {
      "post": {
        "tags": [
          "pauses"
        ],
        "summary": "Cancel Pause",
        "operationId": "cancel_pause_pause_requests__pause_id__cancel_post",
        "parameters": [
          {
            "name": "pause_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Pause Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PauseCancelIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PauseOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cases/{case_id}/escalations": {
      "get": {
        "tags": [
          "escalations"
        ],
        "summary": "List Escalations",
        "operationId": "list_escalations_cases__case_id__escalations_get",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EscalationOut"
                  },
                  "title": "Response List Escalations Cases  Case Id  Escalations Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cases/{case_id}/penalties": {
      "get": {
        "tags": [
          "penalties"
        ],
        "summary": "List Case Penalties",
        "operationId": "list_case_penalties_cases__case_id__penalties_get",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PenaltyOut"
                  },
                  "title": "Response List Case Penalties Cases  Case Id  Penalties Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/penalties/{penalty_id}": {
      "get": {
        "tags": [
          "penalties"
        ],
        "summary": "Get Penalty",
        "operationId": "get_penalty_penalties__penalty_id__get",
        "parameters": [
          {
            "name": "penalty_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Penalty Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PenaltyOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/penalties/{penalty_id}/lock": {
      "post": {
        "tags": [
          "penalties"
        ],
        "summary": "Lock Penalty",
        "operationId": "lock_penalty_penalties__penalty_id__lock_post",
        "parameters": [
          {
            "name": "penalty_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Penalty Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PenaltyOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cases/{case_id}/sla-trace": {
      "get": {
        "tags": [
          "trace"
        ],
        "summary": "Sla Trace",
        "operationId": "sla_trace_cases__case_id__sla_trace_get",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SlaTraceOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CaseCreate": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Title"
          },
          "sla_hours": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Sla Hours",
            "description": "SLA 时限（小时）",
            "default": 48.0
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At",
            "description": "缺省为当前时钟"
          }
        },
        "type": "object",
        "required": [
          "title"
        ],
        "title": "CaseCreate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PauseStatus": {
        "type": "string",
        "enum": [
          "PENDING",
          "APPROVED",
          "REJECTED",
          "ENDED",
          "CANCELLED"
        ],
        "title": "PauseStatus"
      },
...
    }
  }
}